+++ app/src/main/java/com/example/appraisal/UI/main_menu/my_experiment/ExpStatusFragment.java
@@ -20,7 +20,7 @@ import com.example.appraisal.R;
 import com.example.appraisal.UI.main_menu.specific_experiment_details.SpecificExpActivity;
 import com.example.appraisal.backend.experiment.Experiment;
 import com.example.appraisal.backend.user.User;
-import com.example.appraisal.model.MainModel;
+import com.example.appraisal.model.core.MainModel;
 import com.google.firebase.firestore.CollectionReference;
 
 /**
