+++ app/src/main/java/com/example/appraisal/UI/main_menu/my_experiment/MyExperimentActivity.java
@@ -13,7 +13,7 @@ import com.example.appraisal.UI.main_menu.MainMenuCommonActivity;
 import com.example.appraisal.R;
 import com.example.appraisal.backend.experiment.Experiment;
 import com.example.appraisal.backend.user.User;
-import com.example.appraisal.model.MainModel;
+import com.example.appraisal.model.core.MainModel;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.EventListener;
 import com.google.firebase.firestore.FirebaseFirestoreException;
