+++ app/src/main/java/com/example/appraisal/UI/main_menu/specific_experiment_details/details/trial_list/ViewTrialAdapter.java
@@ -1,4 +1,4 @@
-package com.example.appraisal.UI.main_menu.specific_experiment_details;
+package com.example.appraisal.UI.main_menu.specific_experiment_details.details.trial_list;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -8,7 +8,6 @@ import android.widget.ArrayAdapter;
 import android.widget.TextView;
 
 import com.example.appraisal.R;
-import com.example.appraisal.backend.experiment.Experiment;
 import com.example.appraisal.backend.specific_experiment.ViewTrial;
 
 import java.util.ArrayList;
