+++ app/src/main/java/com/example/appraisal/UI/main_menu/specific_experiment_details/participants/SpecificExpContributorsFragment.java
@@ -1,4 +1,4 @@
-package com.example.appraisal.UI.main_menu.specific_experiment_details;
+package com.example.appraisal.UI.main_menu.specific_experiment_details.participants;
 
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -17,8 +17,8 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.appraisal.R;
 import com.example.appraisal.backend.experiment.Experiment;
-import com.example.appraisal.model.MainModel;
-import com.example.appraisal.model.SpecificExpModel;
+import com.example.appraisal.model.core.MainModel;
+import com.example.appraisal.model.main_menu.specific_experiment_details.SpecificExpModel;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.firestore.DocumentReference;
