--- app/src/main/java/com/example/appraisal/model/DataRequestable.java
@@ -1,9 +1,0 @@
-//package com.example.appraisal.model;
-//
-//import com.google.firebase.firestore.CollectionReference;
-//
-//import java.util.List;
-//
-//public interface DataRequestable {
-//
-//}
