--- app/src/main/java/com/example/appraisal/model/MyExperimentModel.java
@@ -1,5 +1,0 @@
-package com.example.appraisal.model;
-
-public class MyExperimentModel {
-
-}
