+++ app/src/main/java/com/example/appraisal/model/main_menu/my_experiment/MyExperimentModel.java
@@ -1,0 +1,5 @@
+package com.example.appraisal.model.main_menu.my_experiment;
+
+public class MyExperimentModel {
+
+}
