+++ app/src/main/java/com/example/appraisal/model/main_menu/specific_experiment_details/SpecificExpModel.java
@@ -1,9 +1,8 @@
-package com.example.appraisal.model;
+package com.example.appraisal.model.main_menu.specific_experiment_details;
 
 import com.example.appraisal.backend.experiment.Experiment;
 import com.example.appraisal.backend.specific_experiment.Quartile;
 import com.example.appraisal.backend.specific_experiment.SpecificExperiment;
-import com.example.appraisal.backend.trial.CountTrial;
 import com.jjoe64.graphview.series.DataPoint;
 
 import java.util.ArrayList;
