+++ javadocs/com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsFragment.html
@@ -102,7 +102,7 @@ var activeTableTab = "activeTableTab";
 <li>androidx.fragment.app.Fragment</li>
 <li>
 <ul class="inheritance">
-<li>com.example.appraisal.UI.main_menu.specific_experiment_details.SpecificExpContributorsFragment</li>
+<li>com.example.appraisal.UI.main_menu.specific_experiment_details.participants.SpecificExpContributorsFragment</li>
 </ul>
 </li>
 </ul>
