+++ javadocs/com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder.html
@@ -102,7 +102,7 @@ var activeTableTab = "activeTableTab";
 <li>androidx.recyclerview.widget.RecyclerView.ViewHolder</li>
 <li>
 <ul class="inheritance">
-<li>com.example.appraisal.UI.main_menu.specific_experiment_details.SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder</li>
+<li>com.example.appraisal.UI.main_menu.specific_experiment_details.participants.SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder</li>
 </ul>
 </li>
 </ul>
