+++ javadocs/com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.html
@@ -102,7 +102,7 @@ var activeTableTab = "activeTableTab";
 <li>androidx.recyclerview.widget.RecyclerView.Adapter&lt;<a href="../../../../../../com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder.html" title="class in com.example.appraisal.UI.main_menu.specific_experiment_details">SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder</a>&gt;</li>
 <li>
 <ul class="inheritance">
-<li>com.example.appraisal.UI.main_menu.specific_experiment_details.SpecificExpContributorsViewAdapter</li>
+<li>com.example.appraisal.UI.main_menu.specific_experiment_details.participants.SpecificExpContributorsViewAdapter</li>
 </ul>
 </li>
 </ul>
@@ -153,7 +153,7 @@ extends androidx.recyclerview.widget.RecyclerView.Adapter&lt;<a href="../../../.
 <th class="colOne" scope="col">Constructor and Description</th>
 </tr>
 <tr class="altColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../../../../../../com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.html#SpecificExpContributorsViewAdapter-android.content.Context-java.util.ArrayList-com.example.appraisal.model.SpecificExpModel-android.content.SharedPreferences-">SpecificExpContributorsViewAdapter</a></span>(android.content.Context&nbsp;context,
+<td class="colOne"><code><span class="memberNameLink"><a href="../../../../../../com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.html#SpecificExpContributorsViewAdapter-android.content.Context-java.util.ArrayList-com.example.appraisal.model.main_menu.specific_experiment_details.SpecificExpModel-android.content.SharedPreferences-">SpecificExpContributorsViewAdapter</a></span>(android.content.Context&nbsp;context,
                                   java.util.ArrayList&lt;java.lang.String&gt;&nbsp;experimenters,
                                   <a href="../../../../../../com/example/appraisal/model/SpecificExpModel.html" title="class in com.example.appraisal.model">SpecificExpModel</a>&nbsp;model,
                                   android.content.SharedPreferences&nbsp;pref)</code>&nbsp;</td>
@@ -179,7 +179,7 @@ extends androidx.recyclerview.widget.RecyclerView.Adapter&lt;<a href="../../../.
 </tr>
 <tr id="i1" class="rowColor">
 <td class="colFirst"><code>void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../../../../com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.html#onBindViewHolder-com.example.appraisal.UI.main_menu.specific_experiment_details.SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder-int-">onBindViewHolder</a></span>(<a href="../../../../../../com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder.html" title="class in com.example.appraisal.UI.main_menu.specific_experiment_details">SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder</a>&nbsp;holder,
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../../../../com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.html#onBindViewHolder-com.example.appraisal.UI.main_menu.specific_experiment_details.participants.SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder-int-">onBindViewHolder</a></span>(<a href="../../../../../../com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder.html" title="class in com.example.appraisal.UI.main_menu.specific_experiment_details">SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder</a>&nbsp;holder,
                 int&nbsp;position)</code>&nbsp;</td>
 </tr>
 <tr id="i2" class="altColor">
@@ -216,7 +216,7 @@ extends androidx.recyclerview.widget.RecyclerView.Adapter&lt;<a href="../../../.
 <!--   -->
 </a>
 <h3>Constructor Detail</h3>
-<a name="SpecificExpContributorsViewAdapter-android.content.Context-java.util.ArrayList-com.example.appraisal.model.SpecificExpModel-android.content.SharedPreferences-">
+<a name="SpecificExpContributorsViewAdapter-android.content.Context-java.util.ArrayList-com.example.appraisal.model.main_menu.specific_experiment_details.SpecificExpModel-android.content.SharedPreferences-">
 <!--   -->
 </a>
 <ul class="blockListLast">
@@ -252,7 +252,7 @@ public&nbsp;<a href="../../../../../../com/example/appraisal/UI/main_menu/specif
 </dl>
 </li>
 </ul>
-<a name="onBindViewHolder-com.example.appraisal.UI.main_menu.specific_experiment_details.SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder-int-">
+<a name="onBindViewHolder-com.example.appraisal.UI.main_menu.specific_experiment_details.participants.SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder-int-">
 <!--   -->
 </a>
 <ul class="blockList">
