+++ javadocs/com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpDataAnalysisFragment.html
@@ -102,7 +102,7 @@ var activeTableTab = "activeTableTab";
 <li>androidx.fragment.app.Fragment</li>
 <li>
 <ul class="inheritance">
-<li>com.example.appraisal.UI.main_menu.specific_experiment_details.SpecificExpDataAnalysisFragment</li>
+<li>com.example.appraisal.UI.main_menu.specific_experiment_details.analysis.SpecificExpDataAnalysisFragment</li>
 </ul>
 </li>
 </ul>
