+++ javadocs/com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpDetailsFragment.html
@@ -102,7 +102,7 @@ var activeTableTab = "activeTableTab";
 <li>androidx.fragment.app.Fragment</li>
 <li>
 <ul class="inheritance">
-<li>com.example.appraisal.UI.main_menu.specific_experiment_details.SpecificExpDetailsFragment</li>
+<li>com.example.appraisal.UI.main_menu.specific_experiment_details.details.SpecificExpDetailsFragment</li>
 </ul>
 </li>
 </ul>
