+++ javadocs/com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpQRCodeFragment.html
@@ -96,7 +96,7 @@
 <li>androidx.fragment.app.Fragment</li>
 <li>
 <ul class="inheritance">
-<li>com.example.appraisal.UI.main_menu.specific_experiment_details.SpecificExpQRCodeFragment</li>
+<li>com.example.appraisal.UI.main_menu.specific_experiment_details.qr.SpecificExpQRCodeFragment</li>
 </ul>
 </li>
 </ul>
