+++ javadocs/com/example/appraisal/UI/main_menu/specific_experiment_details/ViewTrialActivity.html
@@ -123,7 +123,7 @@ var activeTableTab = "activeTableTab";
 <li>androidx.appcompat.app.AppCompatActivity</li>
 <li>
 <ul class="inheritance">
-<li>com.example.appraisal.UI.main_menu.specific_experiment_details.ViewTrialActivity</li>
+<li>com.example.appraisal.UI.main_menu.specific_experiment_details.details.trial_list.ViewTrialActivity</li>
 </ul>
 </li>
 </ul>
