+++ javadocs/com/example/appraisal/UI/main_menu/specific_experiment_details/ViewTrialAdapter.html
@@ -105,7 +105,7 @@ var activeTableTab = "activeTableTab";
 <li>android.widget.ArrayAdapter&lt;<a href="../../../../../../com/example/appraisal/backend/specific_experiment/ViewTrial.html" title="class in com.example.appraisal.backend.specific_experiment">ViewTrial</a>&gt;</li>
 <li>
 <ul class="inheritance">
-<li>com.example.appraisal.UI.main_menu.specific_experiment_details.ViewTrialAdapter</li>
+<li>com.example.appraisal.UI.main_menu.specific_experiment_details.details.trial_list.ViewTrialAdapter</li>
 </ul>
 </li>
 </ul>
