+++ javadocs/com/example/appraisal/backend/user/FirebaseAuthentication.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.example.appraisal.model.FirebaseAuthentication</li>
+<li>com.example.appraisal.model.core.FirebaseAuthentication</li>
 </ul>
 </li>
 </ul>
