+++ javadocs/com/example/appraisal/model/MainModel.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.example.appraisal.model.MainModel</li>
+<li>com.example.appraisal.model.core.MainModel</li>
 </ul>
 </li>
 </ul>
