+++ javadocs/com/example/appraisal/model/MyExperimentModel.html
@@ -93,7 +93,7 @@
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.example.appraisal.model.MyExperimentModel</li>
+<li>com.example.appraisal.model.main_menu.my_experiment.MyExperimentModel</li>
 </ul>
 </li>
 </ul>
