+++ javadocs/com/example/appraisal/model/SpecificExpModel.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.example.appraisal.model.SpecificExpModel</li>
+<li>com.example.appraisal.model.main_menu.specific_experiment_details.SpecificExpModel</li>
 </ul>
 </li>
 </ul>
