+++ javadocs/deprecated-list.html
@@ -87,7 +87,7 @@
 </tr>
 <tbody>
 <tr class="altColor">
-<td class="colOne"><a href="com/example/appraisal/backend/user/FirebaseAuthentication.html#signOut--">com.example.appraisal.model.FirebaseAuthentication.signOut()</a></td>
+<td class="colOne"><a href="com/example/appraisal/backend/user/FirebaseAuthentication.html#signOut--">com.example.appraisal.model.core.FirebaseAuthentication.signOut()</a></td>
 </tr>
 </tbody>
 </table>
