+++ javadocs/index-files/index-13.html
@@ -77,7 +77,7 @@
 <dd>
 <div class="block">This method is called when th DialogFragment is attached to the MyExperimentActivity</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.html#onBindViewHolder-com.example.appraisal.UI.main_menu.specific_experiment_details.SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder-int-">onBindViewHolder(SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder, int)</a></span> - Method in class com.example.appraisal.UI.main_menu.specific_experiment_details.<a href="../com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.html" title="class in com.example.appraisal.UI.main_menu.specific_experiment_details">SpecificExpContributorsViewAdapter</a></dt>
+<dt><span class="memberNameLink"><a href="../com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.html#onBindViewHolder-com.example.appraisal.UI.main_menu.specific_experiment_details.participants.SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder-int-">onBindViewHolder(SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder, int)</a></span> - Method in class com.example.appraisal.UI.main_menu.specific_experiment_details.<a href="../com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.html" title="class in com.example.appraisal.UI.main_menu.specific_experiment_details">SpecificExpContributorsViewAdapter</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../com/example/appraisal/UI/main_menu/forum/ForumHomeActivity.html#onCreate-android.os.Bundle-">onCreate(Bundle)</a></span> - Method in class com.example.appraisal.UI.main_menu.forum.<a href="../com/example/appraisal/UI/main_menu/forum/ForumHomeActivity.html" title="class in com.example.appraisal.UI.main_menu.forum">ForumHomeActivity</a></dt>
 <dd>&nbsp;</dd>
