+++ javadocs/index-files/index-17.html
@@ -191,7 +191,7 @@
 <dd>
 <div class="block">This is an adapter to show the all Users who have added a trial to the experiment (i.e contributed)</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.html#SpecificExpContributorsViewAdapter-android.content.Context-java.util.ArrayList-com.example.appraisal.model.SpecificExpModel-android.content.SharedPreferences-">SpecificExpContributorsViewAdapter(Context, ArrayList&lt;String&gt;, SpecificExpModel, SharedPreferences)</a></span> - Constructor for class com.example.appraisal.UI.main_menu.specific_experiment_details.<a href="../com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.html" title="class in com.example.appraisal.UI.main_menu.specific_experiment_details">SpecificExpContributorsViewAdapter</a></dt>
+<dt><span class="memberNameLink"><a href="../com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.html#SpecificExpContributorsViewAdapter-android.content.Context-java.util.ArrayList-com.example.appraisal.model.main_menu.specific_experiment_details.SpecificExpModel-android.content.SharedPreferences-">SpecificExpContributorsViewAdapter(Context, ArrayList&lt;String&gt;, SpecificExpModel, SharedPreferences)</a></span> - Constructor for class com.example.appraisal.UI.main_menu.specific_experiment_details.<a href="../com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.html" title="class in com.example.appraisal.UI.main_menu.specific_experiment_details">SpecificExpContributorsViewAdapter</a></dt>
 <dd>&nbsp;</dd>
 <dt><a href="../com/example/appraisal/UI/main_menu/specific_experiment_details/SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder.html" title="class in com.example.appraisal.UI.main_menu.specific_experiment_details"><span class="typeNameLink">SpecificExpContributorsViewAdapter.SpecificExpContributorsViewHolder</span></a> - Class in <a href="../com/example/appraisal/UI/main_menu/specific_experiment_details/package-summary.html">com.example.appraisal.UI.main_menu.specific_experiment_details</a></dt>
 <dd>&nbsp;</dd>
